requirements.txt
# Runtime dependencies for Research_FREnTE
# Target interpreter: Python 3.10+

pydantic>=2.7,<3.0
httpx>=0.27,<1.0
PyYAML>=6.0,<7.0
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
firecrawl-py>=1.0,<2.0
